proxy; pending fullMesh client requests can be closed
The proxy functionality that processes NetwotkServiceEndpoint changes
is no longer blocked on establishing connection to new a endpoint.
Pending connection requests towards disappeared endpoints are removed.
So e.g., during Meridio upgrade the terminating NetworkServiceEndpoints
of an old release can't stall operation of freshly started proxies anymore.

-pkg/client moved to cmd/proxy/internal
-clean-up can take place upon proxy shutdown (e.g., closing fullMesh clients,
 unregistering nse)
